Update patch set 2
Patch Set 2:

Matthias came with idea of leaving Requires as they are now and transform them to meta-requires one by one as they will have correct explicit Provides. That would make this patch pass the gate and we could solve the situation 

Regarding running as root. We had basically to options. Either run fluentd as root (eg. rebuild single package) or modify permissions for log directory of all services we are gathering logs from (eg. rebuild tens of packages). Back then I did not find any suggestions that it is not a good idea to run fluentd as root, so we took the easiest from the possible options. We can and will adapt opstools-ansible when this patch will merge. In the future then we will have all required package rebuild with correct permissions on log directories we will go back to running as fluentd user, which is definitely better situation, but now we are deadlocked with upstream spec file out of sync.

Patch-set: 2