Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Yes this works for us, because hostname is installed on CentOS and THEL by default. The case where it does not work is OracleLinux, because the package is not part of clean image. For the reason above I don't think we need a BZ. Just changelog and commit msg should contain clear description, why this patch is needed.

Patch-set: 2
Label: Code-Review=-1