Update patch set 1
Patch Set 1:

I think using jemalloc 3.x is fine.  If someone else is already supporting and providing jemalloc for centos, let's use that, and update the version if necessary.  Who is the maintainer?
The only reason I used jemalloc 5.x is because that was the latest version available at the time.  My goal was to provide jemalloc for _only_ fluentd to use, so we (opstools) would not have to maintain a jemalloc that a lot of teams would use.  I really don't want to take on additional package maintainership burden, especially if it is a general purpose package like jemalloc (although a lot of the rubygem packages we have to maintain for fluentd can also be used by many other ruby applications).

Patch-set: 1
Label: Workflow=0